make example work
